+++ packages/orm/angel_orm_test/lib/src/models/car.g.dart
@@ -91,15 +91,15 @@ class CarQuery extends Query<Car, CarQueryWhere> {
     }
     var model = Car(
         id: fields.contains('id') ? row[0].toString() : null,
-        createdAt: fields.contains('created_at') ? (row[1] as DateTime?) : null,
-        updatedAt: fields.contains('updated_at') ? (row[2] as DateTime?) : null,
+        createdAt: fields.contains('created_at') ? mapToDateTime(row[1]) : null,
+        updatedAt: fields.contains('updated_at') ? mapToDateTime(row[2]) : null,
         make: fields.contains('make') ? (row[3] as String?) : null,
         description:
             fields.contains('description') ? (row[4] as String?) : null,
         familyFriendly:
             fields.contains('family_friendly') ? mapToBool(row[5]) : null,
         recalledAt:
-            fields.contains('recalled_at') ? (row[6] as DateTime?) : null);
+            fields.contains('recalled_at') ? mapToDateTime(row[6]) : null);
     return Optional.of(model);
   }